decode: Don't shift signed integers left
Fixes #3.

+++ decode.c
@@ -2,8 +2,8 @@
 #include "disarm64.h"
 #include <stdint.h>
 
-static int64_t sext(uint64_t imm, unsigned bits) {
-  uint64_t sign = 1 << (bits - 1);
+static uint64_t sext(uint64_t imm, unsigned bits) {
+  uint64_t sign = (uint64_t)1 << (bits - 1);
   return imm & sign ? (imm ^ sign) - sign : imm;
 }
